Auth now specifies api as audience.
partially fixes #2263

In the baseline implementation of the app the auth token request did not
specify an audience (the target API for the JWT). After defining the
policyengine API in auth0, I have now updated both the provider and the
fetch with auth to specify issuer.

Immediately this will actually stop existing, logged in users from
submitting a bearer token until they have to re-log in.

Since the bearer token is yet used this change will not impact the user
experience.